app/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-Bcrypt==0.7.1
PyMySQL==0.9.3
requests==2.22.0
uWSGI==2.0.18
Pillow==6.2.2
pep8
pylint